Skip failing API tests
We currently have a lot of API tests failing due to open bugs (OR)
unimplemented features. This patch updates the API tests to skip the
failing tests. This is necessary to make sure that currently working
code does not break as a result of regression failures. The skipped
tests will need to re-introduced as the open bugs/features are
addressed.

Change-Id: Id667b641c5829683b1478911caa4d072f5706001