Enable local development using fig/docker setup
With this change, you can create a local development environment using fig,
make changes to local code, causing the uwsgi application to be automatically
reloaded in the poppy Docker container.

This change also fixes the repose/init.sh file to account for the new
location of the repose-valve.jar file.

Change-Id: I4e1ac57cecdba717a6f92917103ac778be449cc5