Made the X-Project_ID header mandatory in Poppy
Any middleware used (Keystone, Repose, etc) will be required to
validate that the project_id is correct for the auth token used.

This change also moves the instantiation of the hook that checks
that the X-Project-ID header is mandatory to each controller
instead of during the make_app.  This allows the X-Project-ID
context hook to only be checked on certain controllers (and ignored
on ones like /ping where projectid is not needed.

Change-Id: Ie2cfb9d76a0e53ba990332d1050cce634626b992
Closes-Bug: 1392818