Fixed Create Flavor Tests
Rather than skip all the negative tests, I added the ability to
specify which scenario's in the ddt file to skip.  For now this
is just the limit checks that are not implemented.

The Negative test to check the invalid href was not a valid test
so it has been removed.  Ie the href was still valid.  With unicode
I dont believe there are any invalid hrefs.

Change-Id: Icc8550d63b77ed977001c1f6417bbce888307df4