Added a test for replacing a restrictions property
This change is for adding test data into the data_patch_service.json file
that will add a restriction property and then rename it using the PATCH
API call with the ops sub-property set to replace. This covers a test for
replacing information for a restriction property in a service.

Fixed pep8 issue of too many lines in test_services.py

Jenkins through a pep8 issue of too many lines at line 508.
It was actually at line 539. I removed the extra line.

Change-Id: I763dfe3143212518c5f9490ff3a965a2f0570acd