Fixes an issue where uniqueness is not enforced for rules
This specific scenario handles the case where /* and * are treated
as different rules and thus unique constraint is not enforced.

The service however then fails to deploy as /* and * are treated the
same in the Akamai driver.

This fix treats the preceding slash and lack of preceding slash as the
same during validation, and then leaves it to the individual provider on
how to treat it (ie keep it, add it, or strip it).

Change-Id: I242fe93542fcf6329bd8287e06ee5a268c033a52