Adds the project id to the service object
This patch adds the project id to the service object so that the
project_id of the user can be referenced by other parts of the code.

Examples where it may be needed is when creating services at the
cdn provider driver, by allocating the service to the correct subcustomer.

Change-Id: Ib2816f55549788242eaf1ac60990b2ec06b8eef0