Return the project id in the service
The service now includes the project id in the response when performing
a GET. The advantage of doing this is you dont have to parse the
links section to find the projectid of the user the service belongs to.

This is mostly effective in the admin endpoint when retrieving a service
by domain name in order to identify the customer.

Change-Id: I0df973af1d8b5b2199bfea43b391359523577648