Replaced %s with string formatters PEP 3101
PEP 3101 proposes a new system for built-in string formatting
operations, intended as a replacement for the existing '%' string
formatting operator.

The '%' operator is primarily limited by the fact that it is a
binary operator, and therefore can take at most two arguments.
One of those arguments is already dedicated to the format string,
leaving all other variables to be squeezed into the remaining
argument.

Change-Id: I73cd01bcc4f45dfe1730363481a263b27f3d066e
Closes-Bug: 1423611